Logout added before tests run

+++ frontend/cypress/support/index.js
@@ -18,12 +18,17 @@ import './commands';
 
 it('log out and back in before testing', () => {
   cy.visit('http://localhost:3000');
-  cy.clearCookies();
-  cy.clearLocalStorage();
-  cy.window().then((win) => {
-    win.sessionStorage.clear();
-    win.localStorage.clear();
-  });
+
+  if (document.getElementById('logout') !== null) {
+    cy.get('#logout').click();
+    cy.clearCookies();
+    cy.clearLocalStorage();
+    cy.window().then((win) => {
+      win.sessionStorage.clear();
+      win.localStorage.clear();
+    });
+  }
+
   cy.reload();
 
   if (document.getElementById('username') !== null) {
